+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - added `addItem` and `removeItem` functions to validate dynamically created fields
   - added addConstraint, updateConstraint and removeConstraint API to dynamically
     update fields constraints
+  - `.delete()` method removes now all parsley related classes
 
 ** 1.1.8 **